Upload an additional file to PollyPhi
The new ML Model is now being used to mark peaks as good and bad.
The variables 'markedGoodByCloudModel' and 'markedBadByCloudModel'
are being used keep track of whether the peak was marked as good
or bad by the ML Model. This data along with the names of isotopes
and the associated compounds are being uploaded to PollyPhi in a
separate file