Delete timer used for "No Internet" warning
Whenever EPI is brought up on a system where internet connection is
not available, a warning is shown and a timer starts ticking that
closes the warning (and EPI dialog) automatically after 5 seconds.
This timer however was not being destroyed due to which even when a
connection to the internet has been re-established in that session,
the EPI dialog will not be able to stay for longer than 5 seconds.
This issue has been fixed.